Added inlining; AWSJobStoreTest.test() passed
